Adding errorHandler middleware and modifying service layer

+++ src/restaurants/services/restaurant.service.ts
@@ -1,3 +1,4 @@
+import ExpressReviewsError from "../../utils/error/ExpressReviewsError";
 import * as restaurantDB from "../data/restaurant.data";
 import { Restaurant, RestaurantParams } from "../models/restaurant.model";
 
@@ -33,6 +34,15 @@ export async function createRestaurant(
 ): Promise<Restaurant> {
   const { name, address, category } = data;
   try {
+    const restaurantId = await getRestaurantIdByData(data);
+    if (restaurantId) {
+      throw new ExpressReviewsError(
+        "El restaurante ya existe",
+        500,
+        "ServiceError",
+        Error
+      );
+    }
     return await restaurantDB.createRestaurant(name, address, category);
   } catch (error) {
     throw error;
@@ -44,6 +54,15 @@ export async function updateRestaurant(
   id: number
 ): Promise<Restaurant> {
   try {
+    const restaurant = await getRestaurantById(id);
+    if (!restaurant) {
+      throw new ExpressReviewsError(
+        "El restaurante no existe",
+        500,
+        "ServiceError",
+        Error
+      );
+    }
     return await restaurantDB.updateRestaurant(data, id);
   } catch (error) {
     throw error;
@@ -52,6 +71,15 @@ export async function updateRestaurant(
 
 export async function deleteRestaurant(id: number): Promise<Restaurant> {
   try {
+    const restaurant = await getRestaurantById(id);
+    if (!restaurant) {
+      throw new ExpressReviewsError(
+        "El restaurante no existe",
+        500,
+        "ServiceError",
+        Error
+      );
+    }
     return await restaurantDB.deleteRestaurant(id);
   } catch (error) {
     throw error;
